+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Pixel Banner plugin will be documented in this file.
 ### v3.1.0 - 2025-03-19
 #### ✨ Added
 - Modal for setting banner source from a URL
+- Add voting functionality to Pixel Banner Store
 - Added clear instruction to the top of the "Set Banner Icon" modal
 - Enhance the visibility of the "🌱 GROW YOUR IDEA" button on the AI Banner modal to make it more prominent, highlighting its usefulness.